+++ frontend/src/components/Login/LoginCheck.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const LoginCheck = ({ errorMessage }) => {    
+    return (
+        <div>
+            <div className='text-small-size text-red-500 mb-4'>{errorMessage}</div>
+        </div>
+    );
+};
+
+export default LoginCheck;
